refactor: AccompanyBoardRepository의 메소드를 AccompanyUserRepository로 옮긴다.
