+++ src/main/java/com/dnd/accompany/domain/accompany/infrastructure/AccompanyRequestRepository.java
@@ -1,16 +1,10 @@
 package com.dnd.accompany.domain.accompany.infrastructure;
 
-import java.util.Optional;
-
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.stereotype.Repository;
 
 import com.dnd.accompany.domain.accompany.entity.AccompanyRequest;
-import com.dnd.accompany.domain.accompany.entity.enums.Role;
 import com.dnd.accompany.domain.accompany.infrastructure.querydsl.interfaces.AccompanyRequestRepositoryCustom;
 
-@Repository
 public interface AccompanyRequestRepository extends JpaRepository<AccompanyRequest, Long>, AccompanyRequestRepositoryCustom {
 	void deleteByAccompanyBoardId(Long boardId);
 }
