+++ src/main/java/com/dnd/accompany/domain/accompany/infrastructure/querydsl/interfaces/AccompanyBoardRepositoryCustom.java
@@ -13,8 +13,6 @@ import com.dnd.accompany.domain.accompany.entity.enums.Region;
 public interface AccompanyBoardRepositoryCustom {
 	Slice<FindBoardThumbnailsResult> findBoardThumbnails(Pageable pageable, Region region);
 
-	Optional<UserProfileThumbnail> findUserProfileThumbnail(Long userId);
-
 	boolean isHostOfBoard(Long userId, Long boardId);
 
 	Optional<String> findNickname(Long userId);
